Move code to projects.ts
